# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "UniFi Protect API",
    "description": "Cloudflare Worker backend for UniFi Protect integration with Home Assistant, featuring security sweeps, patrol configurations, and camera management",
    "version": "1.0.0",
    "contact": {
      "name": "UniFi Protect API",
      "url": "https://github.com/jmbish04/hassio-unifi-protect-backend"
    }
  },
  "servers": [
    {
      "url": "https://protect.hacolby.app",
      "description": "Production server"
    },
    {
      "url": "https://unifi-protect-api.{username}.workers.dev",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/logs": {
      "post": {
        "summary": "Store log entries",
        "description": "Store single or multiple log entries in D1 database with 30-day TTL. No authentication required for writing logs.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/LogEntryRequest"
                  },
                  {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/LogEntryRequest"
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Log entry stored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEntriesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid JSON or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to store log entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve log entries",
        "description": "Query log entries with optional filtering. Authentication required.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of entries to return",
            "schema": {
              "type": "integer",
              "default": 100,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of entries to skip",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "description": "Filter by log level",
            "schema": {
              "type": "string",
              "enum": ["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"]
            }
          },
          {
            "name": "logger_name",
            "in": "query",
            "required": false,
            "description": "Filter by logger name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "description": "Filter by start date (ISO format)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "description": "Filter by end date (ISO format)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "request_id",
            "in": "query",
            "required": false,
            "description": "Filter by request ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "correlation_id",
            "in": "query",
            "required": false,
            "description": "Filter by correlation ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log entries retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEntriesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to retrieve log entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}